Reboot new and improved, less fat
It has become clear that the focus of the Beginner Boost needs to be on
tapping into the intrisic power of computing by learning to leverage the
Linux Bash terminal and thereby empower users as young as 12 to master
the UNIX philosophy and get coding on things they come to care about
immediately (from the terminal). Terminal mastery will help them
throughout their lives, no matter what career path they choose ---
especially since only 0.001% of even the tech population has this
foundation.

> "How to make full use of this complex software that immediately boots
> up with your computer, is oddly one of the least taught things."
> (@jaysega)

The rest of that stuff covered is all fat to be cut, including the
following:

* All Web Development, GitHub Pages, Netlify
* Hardware Setup and Linux OS Installation
* Linux Desktop Environment
* Virtual Machine Installation and Setup
* Networking

Dropping Web Dev might be the most shocking. But it is easily learned
and covered way to much already.

These are, of course, valuable things to learn, but they are not in line
with the mission of the Boost, which has always been to Boost learners
in a way that produces the most empowering base upon which to build the
rest of their skills. Some will find the Boost *after* learning
something else and that is fine. But the primary purpose is to get
beginners to become shell natives as soon as possible.